+++ Serenity.Data.Entity/FieldTypes/Field.cs
@@ -28,6 +28,9 @@ namespace Serenity.Data
         internal int naturalOrder;
         internal string textualField;
         private Criteria criteria;
+        internal string readPermission;
+        internal string insertPermission;
+        internal string updatePermission;
 
         protected Field(ICollection<Field> fields, FieldType type, string name, LocalText caption, int size, FieldFlags flags)
         {
@@ -259,6 +262,12 @@ namespace Serenity.Data
             get; set;
         }
 
+        public string InsertPermission
+        {
+            get { return insertPermission; }
+            set { insertPermission = value; }
+        }
+
         public SelectLevel MinSelectLevel
         {
             get { return minSelectLevel; }
@@ -271,10 +280,22 @@ namespace Serenity.Data
             set { naturalOrder = value; }
         }
 
+        public string ReadPermission
+        {
+            get { return readPermission; }
+            set { readPermission = value; }
+        }
+
         public string TextualField
         {
             get { return textualField; }
             set { textualField = value; }
+        }
+
+        public string UpdatePermission
+        {
+            get { return updatePermission; }
+            set { updatePermission = value; }
         }
 
         public LeftJoin ForeignJoin(Int32? foreignIndex = null)