+++ Serenity.Services/RequestHandlers/Delete/DeleteRequestHandler.cs
@@ -181,12 +181,7 @@ namespace Serenity.Services
                 typeof(TRow).GetCustomAttribute<ModifyPermissionAttribute>(false); 
 
             if (attr != null)
-            {
-                if (attr.Permission.IsNullOrEmpty())
-                    Authorization.ValidateLoggedIn();
-                else
-                    Authorization.ValidatePermission(attr.Permission);
-            }
+                Authorization.ValidatePermission(attr.Permission ?? "?");
         }
 
         public TDeleteResponse Process(IUnitOfWork unitOfWork, TDeleteRequest request)