+++ Serenity.Web/Reporting/ReportRegistry.cs
@@ -91,7 +91,7 @@ namespace Serenity.Reporting
                 return false;
 
             foreach (var report in reports)
-                if (Authorization.HasPermission(report.Permission))
+                if (report.Permission == null || Authorization.HasPermission(report.Permission))
                     return true;
 
             return false;
@@ -102,7 +102,6 @@ namespace Serenity.Reporting
             EnsureTypes();
 
             var list = new List<Report>();
-            var permissionService = Dependency.Resolve<IPermissionService>();
 
             foreach (var k in reportsByCategory)
                 if (categoryKey.IsNullOrEmpty() ||
@@ -110,8 +109,7 @@ namespace Serenity.Reporting
                     (k.Key + "/").StartsWith((categoryKey ?? ""), StringComparison.OrdinalIgnoreCase))
                 {
                     foreach (var report in k.Value)
-                        if (report.Permission.IsNullOrEmpty() ||
-                            permissionService.HasPermission(report.Permission))
+                        if (report.Permission == null || Authorization.HasPermission(report.Permission))
                         {
                             list.Add(report);
                         }
@@ -161,8 +159,8 @@ namespace Serenity.Reporting
                 this.Category = new ReportRegistry.Category(category, GetReportCategoryTitle(category));
 
                 attr = type.GetCustomAttributes(typeof(RequiredPermissionAttribute), false);
-                if (attr.Length == 1)
-                    this.Permission = ((RequiredPermissionAttribute)attr[0]).Permission;
+                if (attr.Length > 0)
+                    this.Permission = ((RequiredPermissionAttribute)attr[0]).Permission ?? "?";
             }
         }