+++ Serenity.Web/Scripts/serenity/CoreLib/UI/DataGrid/TreeGridMixin.ts
@@ -43,7 +43,8 @@
 
             if (options.toggleField) {
                 var col = Q.first(dg.getGrid().getColumns(), x => x.field == options.toggleField);
-                col.format = SlickFormatting.treeToggle(() => dg.view, getId, col.format);
+                col.format = SlickFormatting.treeToggle(() => dg.view, getId,
+                    col.format || (ctx => Q.htmlEncode(ctx.value)));
                 col.formatter = SlickHelper.convertToFormatter(col.format);
             }
         }