+++ CHANGELOG.md
@@ -1,4 +1,32 @@
-## 2.5.8 (2016-11-08)
+## 2.6.1 (2016-11-15)
+
+Features:
+  - ability to use original ID column with GridEditorBase instead of "__id"
+
+Bugfixes:
+  - if an item is just readonly in property grid, it should still be serialized
+  - fix tree grid mixin doesn't work when toggle column has no formatter
+
+## 2.6.0 (2016-11-12)
+
+Features:
+  - implemented field level permissions, just add one or more of ReadPermission, ModifyPermission, InsertPermission, UpdatePermission attributes to properties in a row.
+  - added LogicOperatorPermissionService that allows using & (and), | (or) operators in permission checks, e.g. ReadPermission("A&B|C")
+  - ListField, RowField, RowListField types has NotMapped flag by default so no need to add [NotMapped] attribute explicitly
+  - ListField also supports value comparison just like RowListField
+  - added setSelectedKeys method to GridRowSelectionMixin (thanks @estrusco)
+  - added other form in tab with one toolbar sample (thanks @estrusco) [Serene]
+  - added a report page for Northwind, more report samples are on the way [Serene]
+
+## 2.5.9 (2016-11-07)
+
+Features:
+  - expand category when related link is clicked
+
+Bugfixes:
+  - fields in collapsed categories could be focused
+
+## 2.5.8 (2016-11-06)
 
 Features:
   - TreeGridMixin for tree view like grid functionality